api/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1

# Database
sqlalchemy==2.0.29
psycopg2-binary==2.9.9

# Opensearch
opensearch-py==2.6.0

# Env management
python-dotenv==1.0.1

# Optional but useful
httpx==0.27.0          # for async API calls
requests==2.31.0       # for sync API calls
pydantic-settings==2.1.0  # structured config
python-multipart==0.0.6
